pyproject.toml
[project]
name = "softbot-schema-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bsr",
    "fastapi>=0.115.12",
    "langchain>=0.0.27",
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "openai>=1.72.0",
    "pydantic>=2.11.3",
    "pyelastica==0.3.3",
    "python-dotenv>=1.1.0",
    "scipy>=1.10.1",
    "uvicorn>=0.33.0",
]

[tool.uv.sources]
pyelastica = { git = "https://github.com/GazzolaLab/PyElastica", branch = "update-0.3.3" }
bsr = { git = "https://github.com/GazzolaLab/Blender-Soft-Rod" }
